refactor: rename sdk to monitor
