+++ hyperdash/__init__.py
@@ -1,5 +1,6 @@
-from .sdk import monitor
+from .monitor import monitor
 from .jupyter import IPythonMagicsWrapper as IPythonMagicsWrapper
+from .experiment import Experiment
 
 # No-op just to make import nicer
 def monitor_cell():
